Convert sunday route unit test to integration test
